skill calculator: Fix combined items with decimals
Fixes runelite/runelite#3812

+++ runelite-client/src/main/java/net/runelite/client/plugins/skillcalculator/SkillCalculator.java
@@ -247,7 +247,7 @@ class SkillCalculator extends JPanel
 			slot.setText("Lvl. " + action.getLevel() + " (" + formatXPActionString(xp, actionCount, "exp) - "));
 			slot.setAvailable(currentLevel >= action.getLevel());
 			slot.setOverlapping(action.getLevel() < targetLevel);
-			slot.setValue((int) xp);
+			slot.setValue(xp);
 		}
 
 		updateCombinedAction();
